+++ changelog.md
@@ -1,6 +1,7 @@
 # Fake Rate Changelog
 ## v1.1.0 (2024-04-26)
 - Fixed incompatibility with the mod "More Difficulties" by uproxide
+- Fixed incompatibility with the mod "Grandpa Demon" by ItzKiba
 
 ## v1.0.0 (2024-04-24)
 - Initial release